# Optimize HistogramTimeseries and HistogramValue for speed