Optimize HistogramTimeseries and HistogramValue for speed

This change removes unnecessary nesting from BucketOptions and moves them
from HistogramValue to HistogramTimeseries. The later change allows multiple
points in the timeseries to use the same bucket options and significantly
improves performance in that case.

Benchmark in Go demonstrates the following improvement of encoding and decoding
compared to the baseline state:

```
===== Encoded sizes
Encoding                       Uncompressed  Improved        Compressed  Improved
Baseline/Metric/Histogram       13769 bytes  [1.000], gziped  769 bytes  [1.000]
Proposed/Metric/Histogram       13569 bytes  [1.015], gziped  774 bytes  [0.994]

Encoding                       Uncompressed  Improved        Compressed  Improved
Baseline/Metric/MixSeries       109867 bytes  [1.000], gziped 7006 bytes  [1.000]
Proposed/Metric/MixSeries       97867 bytes  [1.123], gziped 6620 bytes  [1.058]

BenchmarkEncode/Baseline/Metric/HistogramOne-8 	      61	  82956741 ns/op
BenchmarkEncode/Proposed/Metric/HistogramOne-8 	      93	  67504571 ns/op

BenchmarkEncode/Baseline/Metric/HistogramSeries-8         	      20	 278817032 ns/op
BenchmarkEncode/Proposed/Metric/HistogramSeries-8         	      36	 167044602 ns/op

BenchmarkDecode/Baseline/Metric/HistogramOne-8            	      31	 193549762 ns/op	106696051 B/op	 2824000 allocs/op
BenchmarkDecode/Proposed/Metric/HistogramOne-8            	      36	 166249091 ns/op	104296031 B/op	 2624000 allocs/op

BenchmarkDecode/Baseline/Metric/HistogramSeries-8         	       9	 592067126 ns/op	287496035 B/op	 7524000 allocs/op
BenchmarkDecode/Proposed/Metric/HistogramSeries-8         	      15	 374523053 ns/op	233896049 B/op	 5324000 allocs/op
```

For single histogram value timeseries it is about 16-18% faster, for timeseries composed
of 5 histogram values it is about 36-40% faster.

Benchmark source code is available at:
https://github.com/tigrannajaryan/exp-otelproto/blob/4c8771fac0543fd8e8d0892a16dcfeb1be43c3aa/encodings/encoding_test.go